Fix race condition in DiagnosticSourceEventSource
If multiple threads all try to write different objects at the same time, PropertySpec.Fetch can manifest a race condition that results in potentially trying to cast one call's object to another call's type.  The fix is to allow for atomically storing the cached data, and using a local to ensure that a calling thread's view is consistent.