Don't pool Http2Connection buffers
Http2Connection keeps these buffers for a really long time; we don't want it to starve the pool.  It also leads to lifetime issues, as the buffers need to be (but currently aren't) released back to the pool, yet work can end up accessing the ArrayBuffer after the connection is logically disposed.