Simplify ProcessorCount tests to fix sporadic failures
We do complicated logic based on a variety of conditions to detect the processor count on Unix; trying to duplicate that logic in the test would essentially just be a copy and paste of the shipping code, which isn't particularly useful.  Until we can do something better, I'm simplifying the tests to just check that we get a valid count out of ProcessorCount.