Console: toggle terminal echo based on presence of interactive child processes
.NET applications echo characters during a Console.Read. By default,
Unix terminals echo characters as the user is typing them.

When a .NET Core application launches an interactive application (e.g. 'vi')
that application expects to find the terminal in an echoing state.

To make both work, corefx was disabling echo during Console.Read operations,
and turning it back on when the read is done.

This changes to echoing while there are interactive applications and not echoing
when there are none.

This means we no longer need to toggle echo off/on for each Console.Read.
And the terminal will no longer echo when there is no Console.Read going on.