Ensure ZipPackage can handle file lengths over int.MaxValue when using FileAccess.Read and FileAccess.Write.
This helps unblock the scenario of using large files with System.IO.Packaging

FileAccess.Write was completely broken before because ZipPackage would create an entry
and immediately delete it, which is disallowed when you're writing an archive.  This
was a porting error, the desktop code had this under an else that was incorrectly removed.

Additionally when reading an archive with a large file, accessing the length of the
stream would cause the entire stream to be read into memory uncompressed and
overflow since this used a MemoryStream backed by a int.max-limited managed array.
Instead use the length info from the archive entry.