[release/2.1] Fix build errors in some build configurations
* Change #if NEED_OPENSSL_1_0 to #ifdef NEED_OPENSSL_1_0

This was a mistake in e4bcbd5885 (#34443) made when backporting it to
the release branches. All other uses of NEED_OPENSSL_1_0 are guarded by
an #ifdef, not an #if. This one should be too.

The warning is seen when building corefx using source-build. I couldn't observe
it directly:

    source-build/src/corefx/src/Native/Unix/System.Security.Cryptography.Native/pal_asn1_print.cpp:34:5:
    error: 'NEED_OPENSSL_1_0' is not defined, evaluates to 0 [-Werror,-Wundef]
      #if NEED_OPENSSL_1_0
      ^

* Disable used-but-marked-unused warnings in System.Security.Cryptography.Native

When building in non-portable mode, some OpenSSL 1.1 function defnitions
that are marked as unused can be picked up by our build. When those
functions are called, clang reports a warning and fails the build:

    src/Native/Unix/System.Security.Cryptography.Native/openssl.cpp:432:12:
    error: 'sk_ASN1_OBJECT_num' was marked unused but was used [-Werror,-Wused-but-marked-unused]
          return sk_ASN1_OBJECT_num(eku);
                 ^

This 'unused' attribute was recently added to sk_* methods in OpenSSL
1.1: https://github.com/openssl/openssl/pull/8246