Improve performance of Pipe.Reader.AsStream()
The Stream returned from the base PipeReader uses public APIs on PipeReader because it needs to work with any PipeReader implementation.  But the vast majority case is a PipeReader returned from Pipe, and the Stream returned from its AsStream can make use of internals on Pipe.  This change does so, reducing the overheads involved in accessing a reader as a stream.