Fix NetEventSource logging for Windows auth on Linux (#35918)
While debugging some Linux enterprise auth (Negotiate/NTLM) scenarios, I discovered that the
logging I added in PR #35383 wasn't working. This was due to not passing in a 'null' value
for the context object for NetEventSource.Info. It was using the string object as the
context object and not logging the string data.

As part of fixing that I optimized some code for remembering what auth package (NTLM
or SPNEGO) is used.