Add SslStream test to validate synchronous invocation on underlying Stream (#36097)
* Change SslStreamStreamToStreamTest_Sync to actually be synchronous

* Add SslStream test to validate synchronous invocation on underlying Stream

SslStream doesn't introduce its own asynchrony.  The only reason an asynchronous operation on SslStream should complete asynchronously is because of an asynchronous call made on its underlying stream completes asynchronously.  This adds tests to validate that, e.g. the tests would fail if we started using Task.Run in the SslStream implementation to invoke operations on the underlying stream.