Fix Chr_CharCodeOutOfRange_ThrowsNotSupportedException VB test
When run in some cultures, the TextInfo.ANSICodePage in place and that's queried by Strings.Chr doesn't fail for the specified inputs, contrary to what's documented (this is true for both .NET Framework and .NET Core).  This change just fixes the test to always use en-US so as to avoid any issues when the test suite is run with other code pages.