Console: Unix: Improve Cursor{Left,Top} performance by caching them (#36049)
* Console: Unix: Improve Cursor{Left,Top} performance by caching them

We keep a cached version for CursorLeft and CursorTop. When we write
data to stdout, we try to update the cached values based on the buffer
content, or we invalidate the cursor position.

A lock is used to deal with multi-threading issues. We synchronize
between some blocks using a 'version' field. This field is incremented
each time a cursor change may have occured. If the value has not
changed between two blocks, we know the the cursor is still at the same
position.

On signals that may mean the cursor position has changed (SIGCONT,
SIGCHLD, SIGWINCH) we invalidate the cached values.

* PR feedback

* lock on Console.Out and rework invalidation on signals

* fix: pass cursorVersion also in UpdateCachedCursorPosition

* Since we check invalidated before reading cached settings, there is no need to invalidate after writing

* Use GetWindowSize instead of getting WindowWidth, WindowHeight

* Update comment

* Fix swapped width and height

* ReadKey: avoid write call for '\0' characters