Avoid thread pool work item allocation in Socket.Send/ReceiveAsync (#36242)
In an async method when you await a task, there are effectively three possible flows:
1. the task is already completed, and execution simply continues
2. the task is not yet completed, we hook up a continuation, and at some point later the task completes and invokes the continuation
3. the task is not yet completed, but by the time we go to hook up the continuation, it has completed.

In case (3), the C# await code is already assuming that things are completing asynchronously, so we can't just keep executing the C# code synchronously.  And if we were to invoke the delegate synchronously, we risk a stack dive, as we're now further down the stack, and if that kept happening, we could run out of stack space.  So we're forced to queue the continuation.

With sockets, case (3) happens more frequently than we might expect: data isn't available immediately but is very soon.  Every time it does, we incur an allocation from using QueueUserWorkItem.

However, in the common case where we're inside an async method, which means it's handling ExecutionContext flow for us, and we can instead use UnsafeQueueUserWorkItem, which has a fast-path that's AsyncStateMachineBox-aware, and can avoid a work item allocation if the continuation/state are coming from an async method.

So, this change just sees whether we need to flow context, and if we don't, uses UnsafeQueueUserWorkItem instead of QueueUserWorkItem, avoiding the allocation in the common case.