ensuring that calling GetPart will call GetPartCore
Package is an abstract class that others can derive from. Those derived classes can override GetPartsCore and potentially not return anything. Furthermore, it is not guaranteed that GetPartsCore will have been called if the package wasn't created using the Package.Open API, which only ensures that ZipPackage's internal data structures are filled in.