Add DataTableExtensions.AsDataView
This pushes down the minimal number of types required to implement
DataTableExtensions.AsDataView.

LinqDataView is added in order to implement AsDataView, the following
types need to be pushed down in order to implement this:
 - DataSetUtil
 - DataTableExtensions
 - EnumerableRowCollection
 - SortExpressionBuilder

Moving thse down requires that the following also be moved due to needing
internals access to moved types.
 - EnumerableRowCollectionExtensions
 - OrderedEnumerableRowCollection
 - TypedTableBase
 - TypedTableBaseExtensions