Add DataTableExtensions.AsDataView (#36528)
* Move most DataSetExtensions code to common

* Add DataTableExtensions.AsDataView

This pushes down the minimal number of types required to implement
DataTableExtensions.AsDataView.

LinqDataView is added in order to implement AsDataView, the following
types need to be pushed down in order to implement this:
 - DataSetUtil
 - DataTableExtensions
 - EnumerableRowCollection
 - SortExpressionBuilder

Moving thse down requires that the following also be moved due to needing
internals access to moved types.
 - EnumerableRowCollectionExtensions
 - OrderedEnumerableRowCollection
 - TypedTableBase
 - TypedTableBaseExtensions

* Add license to LinqDataView.cs

* Add a couple simple AsDataView tests

* Move remaining types from DataSetExtensions to Common

* Clean up source. CR feedback.

* AsDataView test only NetCoreApp

* Make System.Data.DataSetExtensions inbox, remove package

We pushed all the types down into System.Data.Common, and the facade is
now inbox, so the package only remains meaningful to carry a netstandard2.0
implementation assembly.  This assembly is only applicable on older
netcoreapp versions, so rather than ship it from master we'll  service
it from the release/2.1 branch.