Update zlib-intel to v1.2.11.1_jtkv6.3 (#36795)
* inflate: handle windowBits == 16

* deflate_medium: add dist -1 to hash even for long matches

This fixes an issue where a repeat sequence longer than 258 would be
encoded using longer distance values after the first match.

* deflate_medium: avoid emitting a suboptimal literal in the restart case

When we load new data into the window, we invalidate the next match, in
case the match would improve. In this case, the hash has already been
updated with this data, so when we look for a new match it will point
it back at itself. As a result, a literal is generated even when a
better match is available.

This avoids that by catching this case and ensuring we're looking at the
past.