Improve WebClient upload tests stability (#36828)
This PR is a follow-up to PR #36801 which improved HttpClient tests stability.

Some of the WebClient tests were failing with WebExceptions being caused by the
Azure remote endpoint returning 500 status code. Similar to the fixes done to
the HttpClient tests, this PR updates some of the WebClient upload tests to use
the VerifyUpload endpoint which doesn't have to echo back the request body payload.

Some of the WebClient tests upload more complex multi-part formats and can't be
easily changed to use the VerifyUpload endpoint. But those tests aren't using large
data sets so it should be ok.

Closes #36226