Cleanup SocketOptionNameTest
One of the oldest issues #4887 was about the Windows socket option, ReuseUnicastPort.
This test was disabled several years ago because it was very difficult to test in CI since it
required not only a newer version of Windows but also specific network settings on the machine.

At this point, there is not any way to reliably test this. And the end-to-end test isn't really
of high value for .NET Core since we would really be mostly testing an OS feature.

As part of cleaning that issue up, I cleaned up the rest of the tests, some of which were marked
as Outerloop but didn't need to be.

Closes #29929
Closes #21327
Closes #4887