Directory.Move behaves differently on Unix vs. Windows (#37121)
* Directory.Move behaves differently on Unix vs. Windows

We will now throw an exception in Unix platforms like in Windows, if the destination path already exists.

* Modifying ExistingDirectory test method that verifies Move overwrite so its called by all platforms.

We already had a test method that tests Move when the destination path exists. I simply removed the platform attribute that restricted it to Windows and renamed the test method so its universal (as opposed to platform specific).